dpu_poc: only do the search on dpu
Some value to compute the score need to be aggregate from the whole
search, so we need the result from the search from every
segments (every thread of every dpus).
Let's do only the search on dpu and give all the value need to compute
the score to the host.
The host needs to aggregate the concerned value from each
segments (not done yet as the implementation is mono dpu).